Projetos Python/PacoteProcessamentoImagem/requirements.txt
pillow
numpy
